LayoutTests/inspector/animation/nameChanged-expected.txt

Tests for the Animation.nameChanged event.


== Running test suite: Animation.nameChanged
-- Running test case: Animation.nameChanged.NewString
PASS: Animation should have the name 'OldName'.
Changing name...

PASS: Animation should have the name 'NewName'.
